Structure switch: Now use Prawn as the basis.
While LaTeX is a wonderful tool for editing documents directly,
it is not really suitable for being auto-driven from a library.
Users may not understand the errors it displays, or may not
even be able to install it correctly. Prawn on the other side is
a great library, which can easily be installed via RubyGems as
a dependency. This commit is far from feature-complete, but the
basic structure, i.e. the combination of generator and formatter,
already works.